updated error reporting for var reassignment
